requirements.txt
Flask
groq
faiss-cpu
sentence-transformers
python-docx
python-dotenv
numpy